Removing the `isDefault` Group block attribute. It's a fragile solution to keep track of whether a variation has been selected since it adds a potentially useless attribute. In this iteration, we'll keep track of whether to show the placeholder via useState in the editor, and, more permanently, whether the block has innerblocks.
